shedule_api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-decouple
psycopg2-binary
django-cors-headers
coverage